natsort.c: fix compile error: type mismatch
Use a size_t instead of an int to silence the compiler.

+++ natsort.c
@@ -98,7 +98,7 @@ int main(int argc, char **argv)
      char **list = 0;
      int linelen = 0, i;
      int c, opt_ind;
-     int bufsize;
+     size_t bufsize;
 
      static struct option long_options[] = {
 	  { "verbose", 0, NULL, 'v'},
